Added Isomap implementation with supporting files
Attempted to calculate good choice of hyperparameters for Isomap, but our dataset may be too large for the standard Isomap implementation to compute, as it relies on running Floyd-Warshall on the graph formed by letting the distances serve as weights between all n observations.  As we all recall, Floyd-Warshall is O(n^3), which is extraordinarily large for our dataset.  Will examine other methods later.