Added support for KNN, as well as working with image data
Managed to improve the accuracy of working with the image data to around 0.4 validation error by converting each image to grayscale, then working with a histogram of the grayscale values.  The SVM performed the worst on this feature set, with a validation error of around 0.5511.  RandomForest did the best, achieving a validation error of 0.4016.  KNN managed to achieve a minimum validation error of 0.4162 with 31 neighbors involved.